Split tensor constructor accuracy tests
