+++ tests/test_arange.py
@@ -4,30 +4,14 @@ import torch
 import flag_gems
 
 from . import accuracy_utils as utils
-from . import conftest as cfg
 
 
 @pytest.mark.arange
-@pytest.mark.parametrize("start", utils.ARANGE_START)
-@pytest.mark.parametrize("step", [1, 2, 5])
-@pytest.mark.parametrize("end", [128, 256, 1024])
-@pytest.mark.parametrize("dtype", utils.FLOAT_DTYPES + utils.ALL_INT_DTYPES + [None])
-@pytest.mark.parametrize("device", [flag_gems.device, None])
-@pytest.mark.parametrize(
-    "pin_memory", [False, None]
-)  # Since triton only target to GPU, pin_memory only used in CPU tensors.
-def test_arange(start, step, end, dtype, device, pin_memory):
-    ref_out = torch.arange(
-        start,
-        end,
-        step,
-        dtype=dtype,
-        device="cpu" if cfg.TO_CPU else device,
-        pin_memory=pin_memory,
-    )
+@pytest.mark.parametrize("end", [10, 100, 1000, 5.0])
+@pytest.mark.parametrize("dtype", [torch.float32, torch.float16, torch.int64])
+def test_arange(end, dtype):
     with flag_gems.use_gems():
-        res_out = torch.arange(
-            start, end, step, dtype=dtype, device=device, pin_memory=pin_memory
-        )
+        res_out = torch.arange(end, dtype=dtype, device=flag_gems.device)
+    ref_out = torch.arange(end, dtype=dtype, device="cpu")
 
-    utils.gems_assert_equal(res_out, ref_out)
+    utils.gems_assert_equal(res_out.cpu(), ref_out)
